requirements.txt
openai==1.30.1
streamlit==1.35.0
sentence-transformers==2.7.0
tenacity==8.2.3
# PyYAML==6.0.1
aiohttp==3.9.5
rich==13.7.1
httpx==0.27.0
python-dotenv==1.0.1
pyyaml==6.0
plotly==5.22.0